requirements.txt
python-telegram-bot==21.0
python-dotenv==1.0.1
requests==2.32.0

imaplib2==3.6
gspread==6.0.0

Flask==3.0.0
gunicorn==21.2.0
oauth2client==4.1.3







